Simplify runtime validation: replace generator with direct template
Replace validate_tracing_runtime_template.py (generator script) with
validate_agent_tracing.py (direct template script).

**Rationale:**
- Simpler workflow: edit template directly instead of generate → edit → run
- More intuitive: template script has clear TODO markers to fill in
- Reduced complexity: one script instead of two-step generation process

**Changes:**
- Deleted: scripts/validate_tracing_runtime_template.py (generator)
- Added: scripts/validate_agent_tracing.py (direct template)
  - TODO markers for module import, dependencies, function call
  - Checks environment variables (MLFLOW_TRACKING_URI, MLFLOW_EXPERIMENT_ID)
  - Complete validation logic (trace capture, hierarchy, session ID)

**SKILL.md Updates:**
- Updated description: "edit directly before running" instead of "generator"
- Updated CLI reference: "edit TODOs first, then run" instead of generator args
- Simplified workflow documentation

**Usage:**
Before: python validate_tracing_runtime_template.py --module X --entry-point Y → edit output → run
After: Edit scripts/validate_agent_tracing.py (fill TODOs) → run directly
