--- agent-evaluation/scripts/validate_tracing_runtime_template.py
@@ -1,295 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Generate a template script for validating MLflow tracing at runtime.
-
-This script creates a customized Python script that you can customize
-with agent-specific dependencies and configuration.
-
-Usage:
-    python validate_tracing_runtime_template.py                              # Auto-detect
-    python validate_tracing_runtime_template.py --module my_agent.agent      # Specify module
-    python validate_tracing_runtime_template.py --entry-point run_agent      # Specify entry point
-    python validate_tracing_runtime_template.py --module my_agent --entry-point run_agent --autolog-file src/my_agent/__init__.py
-"""
-
-import argparse
-import os
-import sys
-
-from utils import validate_env_vars
-
-
-def generate_validation_code(
-    tracking_uri: str,
-    experiment_id: str,
-    module_name: str,
-    entry_point: str,
-    autolog_file: str | None = None
-) -> str:
-    """Generate Python code for runtime validation.
-
-    Args:
-        tracking_uri: MLflow tracking URI
-        experiment_id: MLflow experiment ID
-        module_name: Agent module name (e.g., "my_agent.agent")
-        entry_point: Entry point function name (e.g., "run_agent")
-        autolog_file: Optional path to file with autolog() call
-
-    Returns:
-        Generated Python script as string
-    """
-
-    autolog_section = ""
-    if autolog_file:
-        autolog_section = f"""
-# Autolog file: {autolog_file}
-# Verify that mlflow.autolog() is called in {autolog_file} before agent initialization
-"""
-
-    return f'''#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-Validate MLflow tracing for {module_name}.{entry_point}
-
-Generated by validate_tracing_runtime_template.py
-"""
-
-import os
-import sys
-import mlflow
-from mlflow import MlflowClient
-
-# Set environment variables
-os.environ["MLFLOW_TRACKING_URI"] = "{tracking_uri}"
-os.environ["MLFLOW_EXPERIMENT_ID"] = "{experiment_id}"
-{autolog_section}
-# Import agent
-from {module_name} import {entry_point}
-
-# Configuration
-TEST_QUERY = "What is MLflow?"
-TEST_SESSION_ID = "test-session-123"
-
-print("=" * 60)
-print("MLflow Tracing Validation")
-print("=" * 60)
-print()
-
-# TODO: Configure your agent's dependencies here
-# IMPORTANT: Update this section with your agent's setup code before running!
-# Example:
-# from your_agent.llm import LLMConfig, LLMProvider
-# llm_config = LLMConfig(model="gpt-4", temperature=0.0)
-# llm_provider = LLMProvider(config=llm_config)
-
-print("⚠  IMPORTANT: Configure your agent's dependencies above before running!")
-print()
-
-# Run test query
-print("Running test query...")
-print(f"  Query: {{TEST_QUERY}}")
-print(f"  Session ID: {{TEST_SESSION_ID}}")
-print()
-
-try:
-    # TODO: Adjust the function call to match your agent's signature
-    # Examples:
-    #   response = {entry_point}(TEST_QUERY, llm_provider)
-    #   response = {entry_point}(TEST_QUERY, session_id=TEST_SESSION_ID)
-    #   response = {entry_point}(TEST_QUERY, llm=llm_provider, session_id=TEST_SESSION_ID)
-    #   response = {entry_point}(TEST_QUERY)
-
-    response = {entry_point}(TEST_QUERY)  # <-- UPDATE THIS LINE
-
-    print("✓ Agent executed successfully")
-    print()
-
-    # Capture trace
-    trace_id = mlflow.get_last_active_trace_id()
-    if not trace_id:
-        print("✗ FAILED: No trace ID captured!")
-        print("  Check that mlflow.autolog() is called before agent execution")
-        sys.exit(1)
-
-    print(f"✓ Trace captured: {{trace_id}}")
-
-    # Get trace details
-    client = MlflowClient()
-    trace = client.get_trace(trace_id)
-
-    # Verify trace structure
-    print()
-    print("Verifying trace structure...")
-
-    if not trace.data.spans:
-        print("✗ FAILED: No spans found in trace")
-        sys.exit(1)
-
-    print(f"✓ Top-level span: {{trace.data.spans[0].name}} ({{trace.data.spans[0].span_type}})")
-
-    # Count total spans (including nested)
-    def count_spans(spans):
-        count = len(spans)
-        for span in spans:
-            if hasattr(span, 'spans') and span.spans:
-                count += count_spans(span.spans)
-        return count
-
-    total_spans = count_spans(trace.data.spans)
-    print(f"✓ Total spans: {{total_spans}}")
-
-    if total_spans < 2:
-        print("⚠  WARNING: Only 1 span found - autolog may not be working")
-        print("  Expected: @mlflow.trace decorator span + autolog library spans")
-    else:
-        print("✓ Multiple spans detected - autolog appears to be working")
-
-    # Print trace hierarchy
-    def print_hierarchy(spans, indent=0):
-        for span in spans:
-            prefix = "    " + "  " * indent
-            print(f"{{prefix}}- {{span.name}} ({{span.span_type}})")
-            if hasattr(span, 'spans') and span.spans:
-                print_hierarchy(span.spans, indent + 1)
-
-    print()
-    print("  Trace hierarchy:")
-    print_hierarchy(trace.data.spans)
-
-    # Check session ID (optional)
-    if "session_id" in trace.info.tags:
-        actual_session_id = trace.info.tags["session_id"]
-        print()
-        if actual_session_id == TEST_SESSION_ID:
-            print(f"✓ Session ID tagged: {{actual_session_id}}")
-        else:
-            print(f"⚠  Session ID mismatch: expected {{TEST_SESSION_ID}}, got {{actual_session_id}}")
-    else:
-        print()
-        print("  ℹ  Note: No session_id tag found (optional for single-turn agents)")
-
-    print()
-    print("=" * 60)
-    print("✓ VALIDATION PASSED")
-    print("=" * 60)
-    print()
-    print("Your agent is properly integrated with MLflow tracing!")
-    print()
-
-except Exception as e:
-    print(f"✗ FAILED: {{str(e)}}")
-    import traceback
-    traceback.print_exc()
-    sys.exit(1)
-'''
-
-
-def main():
-    """Main workflow."""
-    # Parse command-line arguments
-    parser = argparse.ArgumentParser(
-        description="Generate validation template script for MLflow tracing",
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-    )
-    parser.add_argument("--module", help="Agent module name (e.g., 'my_agent.agent')")
-    parser.add_argument("--entry-point", help="Entry point function name (e.g., 'run_agent')")
-    parser.add_argument("--autolog-file", help="File containing autolog() call (e.g., 'src/agent/__init__.py')")
-    parser.add_argument("--output", default="validate_agent_tracing.py", help="Output file name")
-    args = parser.parse_args()
-
-    print("=" * 60)
-    print("MLflow Tracing Validation Template Generator")
-    print("=" * 60)
-    print()
-
-    # Check environment
-    errors = validate_env_vars()
-    if errors:
-        print("✗ Environment validation failed:")
-        for error in errors:
-            print(f"  - {error}")
-        print("\nRun scripts/setup_mlflow.py first")
-        sys.exit(1)
-
-    tracking_uri = os.getenv("MLFLOW_TRACKING_URI")
-    experiment_id = os.getenv("MLFLOW_EXPERIMENT_ID")
-
-    print(f"Tracking URI: {tracking_uri}")
-    print(f"Experiment ID: {experiment_id}")
-    print()
-
-    # Get agent module (must be specified manually)
-    print("Agent module configuration...")
-    agent_module = args.module
-    if not agent_module:
-        print("  ✗ Agent module not specified")
-        print("  Use --module to specify your agent module")
-        print("  Example: --module my_agent.agent")
-        print("\n  To find your agent module:")
-        print("    grep -r 'def.*agent' . --include='*.py'")
-        sys.exit(1)
-    else:
-        print(f"  ✓ Using specified: {agent_module}")
-
-    # Get entry point (must be specified manually)
-    print("\nEntry point configuration...")
-    entry_point = args.entry_point
-    if not entry_point:
-        print("  ✗ Entry point not specified")
-        print("  Use --entry-point to specify your agent's main function")
-        print("  Example: --entry-point run_agent")
-        print("\n  To find entry points with @mlflow.trace:")
-        print("    grep -r '@mlflow.trace' . --include='*.py'")
-        sys.exit(1)
-    else:
-        print(f"  ✓ Using specified: {entry_point}")
-
-    # Get autolog file (optional)
-    if args.autolog_file:
-        print(f"\n✓ Autolog file: {args.autolog_file}")
-    else:
-        print("\n  ℹ  No autolog file specified (optional)")
-        print("  Use --autolog-file if you want to document the autolog location")
-
-    # Generate code
-    print("\n" + "=" * 60)
-    print("Generating Validation Template Script")
-    print("=" * 60)
-
-    code = generate_validation_code(
-        tracking_uri, experiment_id, agent_module, entry_point, args.autolog_file
-    )
-
-    # Write to file
-    output_file = args.output
-    with open(output_file, "w") as f:
-        f.write(code)
-
-    print(f"\n✓ Script generated: {output_file}")
-    print()
-
-    # Make executable
-    try:
-        os.chmod(output_file, 0o755)
-        print(f"✓ Made executable: chmod +x {output_file}")
-    except Exception:
-        pass
-
-    print()
-    print("=" * 60)
-    print("Next Steps")
-    print("=" * 60)
-    print()
-    print(f"1. Review the generated script: {output_file}")
-    print("2. Update the TODO sections with your agent's setup code:")
-    print("   - Add imports for your agent's dependencies")
-    print("   - Configure LLM providers, API keys, or other dependencies")
-    print("   - Update the function call to match your agent's signature")
-    print(f"3. Execute it: python {output_file}")
-    print("4. Verify all validation checks pass")
-    print()
-    print("=" * 60)
-
-
-if __name__ == "__main__":
-    main()
